requirements.txt
Django==3.2.9
django-taggit==1.5.1
django-tinymce==3.3.0
Pillow==8.4.0
PyMySQL==1.0.2

